Product Service Permission added

+++ src/main/java/ecommerce/service/ProductService.java
@@ -24,16 +24,27 @@ public class ProductService {
     }
 
     // ADD PRODUCT
-    public boolean add(String name, String category, double price, int quantity) {
+    public boolean add(User user, String name, String category, double price, int quantity) {
+
+        // PERMISSION
+        if(user.getRole() == Role.CLIENT) return false;
+
         // doesn't create product if wrong parameters or product already existed
         if (name == null || category == null || price < 0 || quantity < 0 || !(productRepo.getProduct(name) == null)) {
             return false;
         }
+
+        // add new product to store
         return productRepo.add(new Product(name, category, price, quantity));
     }
 
     // UPDATE PRODUCT
-    public boolean update(String name, String new_category, Double new_price, Integer new_quantity) {
+    public boolean update(User user, String name, String new_category, Double new_price, Integer new_quantity) {
+
+        // PERMISSION
+        if(user.getRole() == Role.CLIENT) return false;
+
+        // local field
         Product product = productRepo.getProduct(name);
 
         // doesn't update unavailable product
@@ -43,31 +54,51 @@ public class ProductService {
         if (new_category != null) product.setCategory(new_category);
         if (new_price != null && new_price >= 0) product.setPrice(new_price);
         if (new_quantity != null && new_quantity >= 0) product.setQuantity(new_quantity);
-
         return true;
     }
 
     // REMOVE PRODUCT
-    public boolean remove(String name) {
+    public boolean remove(User user, String name) {
+
+        // PERMISSION
+        if(user.getRole() == Role.CLIENT) return false;
+
+        // remove product from store
         return productRepo.remove(name);
     }
 
     // SEARCH FOR PRODUCT
-    public Product search(String name) {
+    public Product search(User user, String name) {
+
+        // PERMISSION
+        if(user.getRole() == Role.CLIENT) return null;
+
+        // search product by string text
         return productRepo.getProduct(name);
     }
 
     // SORT PRICE
-    public List<Product> sortPrice() {
+    public List<Product> sortPrice(User user) {
+
+        // PERMISSION
+        if(user.getRole() == Role.CLIENT) return null;
+
+        // sort catalog by price
         List<Product> catalog = productRepo.getAll();
         catalog.sort(SORT_BY_PRICE);
         return catalog;
     }
 
     // VIEW PRODUCT CATALOG
-    public List<Product> viewAll(int view) {
+    public List<Product> viewAll(User user, int view) {
+
+        // PERMISSION
+        if(user.getRole() == Role.CLIENT) return null;
+
+        // local field
         List<Product> catalog = productRepo.getAll();
 
+        // choose view
         switch(view) {
             case 0 -> catalog.sort(SORT_BY_CATEGORY);
             case 1 -> catalog.sort(SORT_BY_NAME);
@@ -78,14 +109,24 @@ public class ProductService {
     }
 
     // VIEW STORE
-    public List<Product> viewAvailable() {
+    public List<Product> viewAvailable(User user) {
+
+        // PERMISSION
+        if(user.getRole() == Role.ADMIN) return null;
+
+        // filter only available stocks
         return productRepo.getAll().stream()
                 .filter(product -> product.getQuantity() > 0)
                 .toList();
     }
 
     // FILTER PRICE BY BUDGET
-    public List<Product> filterPrice(double max_price) {
+    public List<Product> filterPrice(User user, double max_price) {
+
+        // PERMISSION
+        if(user.getRole() == Role.ADMIN) return null;
+
+        // filter only products under price range
         return productRepo.getAll().stream()
                 .filter(product -> product.getQuantity() > 0)
                 .filter(product -> product.getPrice() < max_price)
@@ -94,12 +135,21 @@ public class ProductService {
 
     // CHECKOUT CLIENT CART
     public boolean checkout(User user, String tax) {
+
+        // PERMISSION
+        if(user.getRole() == Role.ADMIN) return false;
+
+        // local field
         Cart cart = cartRepo.getCart(user.getId());
 
+        // cannot checkout an unavailable cart or an empty cart
         if(cart == null || cart.getProductList().isEmpty()) {
             return false;
         }
+
+        // checkout process
         else {
+
             // total price
             double price = 0.00;
 
@@ -113,6 +163,7 @@ public class ProductService {
                 // reduce product quantity from stock
                 product.setQuantity(product.getQuantity()-item.getValue());
             }
+
             // calculate tax
             price *= (1 + (Double.parseDouble(tax)/100));
 
@@ -125,6 +176,11 @@ public class ProductService {
 
     // GET CLIENT ORDER HISTORY
     public List<Order> orderHistory(User user) {
+
+        // PERMISSION
+        if(user.getRole() == Role.ADMIN) return null;
+
+        // grab all previously placed orders
         return orderRepo.getHistory(user.getId());
     }
 }
